+++ CHANGELOG.md
@@ -91,6 +91,12 @@ different versioning scheme, following the Haskell community's
   is configured with `EpoxyTransportBuilder.SetResolver`.
 * Bond-generated Errors now give clients opaque GUIDs. These GUIDs can be
   matched against emitted metrics for debugging.
+* Epoxy can now be configured to use TLS to secure the connection.
+  * TLS configuration is set via
+    `EpoxyTransportBuilder.SetClientTlsConfig`/`EpoxyTransportBuilder.SetServerTlsConfig`.
+  * See the
+    [TLS example](https://github.com/Microsoft/bond/tree/master/examples/cs/comm/tls/)
+    for even more details.
 
 ## 4.2.1: 2016-06-02 ##